[Functions] reorganize the context hierarchy for functions (#10631)
### Motivation

Currently the context relationship for function, source and sink is not well defined. This prevents some common features to be added once for all and creates some confusion, code duplication in the current repo. As demonstrated in the following graph, this PR changes the hierarchy from left to right. By introducing a common base context, it help solving some issues we are seeing. The base context provides common access to pulsar cluster, state, metrics, and meta-data to make sure all components can reuse it.

![context hierarchy](https://user-images.githubusercontent.com/16407807/118730483-8ebf5200-b7ec-11eb-9220-d41261f148bb.png)



### Modifications

- Remove `ConnectorContext` interface.
- Introduce a `BaseContext` interface. 
- Update existing `Context`, `SourceContext`, `SinkContext` interface to extend the new common interface.
